requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt '.\requirements.in'
#
asgiref==3.8.1
    # via django
build==1.2.1
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
distlib==0.3.8
    # via virtualenv
django==5.0.6
    # via
    #   -r .\requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r .\requirements.in
filelock==3.15.4
    # via virtualenv
identify==2.6.0
    # via pre-commit
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r .\requirements.in
platformdirs==4.2.2
    # via virtualenv
pre-commit==3.7.1
    # via -r .\requirements.in
psycopg2==2.9.9
    # via -r .\requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r .\requirements.in
pyyaml==6.0.1
    # via pre-commit
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
uuid==1.30
    # via -r .\requirements.in
virtualenv==20.26.3
    # via pre-commit
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
